Clear selection via DEL/BACKSPACE
